@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -32,7 +32,8 @@ dependencies:
   animated_splash_screen: ^1.1.0
   lottie: ^1.0.1
   blur: ^2.0.0
-  dotted_border: ^1.0.7
+  dotted_border: ^2.0.0
+  
   
 dev_dependencies:
   flutter_test:
